Put methods into region (lmao is so much code)
Unironically rip to Edward for having to grade this code, am trying to keep it tidy, is abit rough. but theres just so much. maybe need to separate in different classes?
Will probably be a 1000+ line fever dream of a C# class by the end of this

I swear i try mine best